fix our UTF-8 output
We mark STDIN/STDOUT as UTF-8, but when we e.g. spew out the Perl
program we want to just spew out raw UTF-8 octets.

We could spew out UTF-8 escaped strings and just set binmode :utf8
there, but I'm trying to keep operations to a minimum.

In the C and Lua output on the other hand we *do* want our strings to
be UTF-8 marked, since we're not escaping them with some dump module,
but our own custom code.